+++ src/EndPoints/Languages.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace seregazhuk\HeadHunterApi\EndPoints;
+
+use seregazhuk\HeadHunterApi\Traits\HasAll;
+
+class Languages extends Endpoint
+{
+    use HasAll;
+
+    const RESOURCE = 'languages';
+}