+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
  - Doesn't throw exception on http errors (4xx and 5xx statuses)
 
 ### Added:
- - Metro, Auth
+ - Metro, Auth, Languages
 
 ## v0.2.6 - 2017-05-05
 ### Changed: